requirements.txt
streamlit==1.31.0
transformers==4.37.2
torch==2.1.2
nltk==3.8.1
pyttsx3==2.90
SpeechRecognition==3.10.0
Pillow==10.1.0
opencv-python==4.8.1.78
matplotlib==3.8.2
pandas==2.1.4
langchain==0.1.0
python-dotenv==1.0.0
huggingface-hub==0.20.2
pytesseract==0.3.10
soundfile==0.12.1
sentencepiece==0.1.99
librosa==0.10.1
protobuf==4.25.1
accelerate==0.25.0
bitsandbytes==0.41.1 